requirements.txt
# Core pipeline (ingestion, features, training)
numpy>=1.24
pandas>=2.0
lxml>=5.0
pyarrow>=14.0
pyyaml>=6.0
python-dotenv>=1.0
finnhub-python>=2.4
yfinance>=0.2.40
huggingface_hub>=0.24
xgboost>=2.0
scikit-learn>=1.3
joblib>=1.3
shap>=0.44

# Dev
pytest>=8.0
